+++ CHANGELOG.md
@@ -35,6 +35,10 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- Check for empty messages improved
+
 ## [2.10.0] - 2025-01-30
 
 ### Added